+++ packages/vuetify/src/components/VInput/index.ts
@@ -1,4 +1,1 @@
-import VInput from './VInput'
-
-export { VInput }
-export default VInput
+export { default as VInput } from './VInput'
