+++ packages/vuetify/src/components/index.ts
@@ -36,7 +36,7 @@ export * from './VGrid'
 // export * from './VHover'
 export * from './VIcon'
 export * from './VImg'
-// export * from './VInput'
+export * from './VInput'
 // export * from './VItemGroup'
 export * from './VKbd'
 export * from './VItemGroup'
@@ -73,7 +73,7 @@ export * from './VResponsive'
 export * from './VSystemBar'
 // export * from './VTabs'
 // export * from './VTextarea'
-// export * from './VTextField'
+export * from './VTextField'
 export * from './VThemeProvider'
 export * from './VTimeline'
 // export * from './VTimePicker'
